Bug fix in output.

+++ IoData.h
@@ -1586,7 +1586,7 @@ struct ROB {
   int numROB;
   int numROBJac;
   int numROBRes;
-  int sampleNodeFactor;
+  double sampleNodeFactor;
   enum ROMSolver {PG = 0, BROYDENPG = 1, GAPPYPG = 2, OLDGAPPYPG = 3} romsolver;
 
   ROB();
